fix: call undecleared waitpid function on linux
When attempting to build a project on Debian Bookworm (unstable), an error is encountered in the file `flutter_pty/src/flutter_pty_unix.c` at line 114. The error message indicates that the function `waitpid` is undeclared and that ISO C99 and later versions do not support implicit function declarations.

To resolve this issue, you can include the header file `<sys/wait.h>` in the `flutter_pty_unix.c` file. This header file provides the declaration for the `waitpid` function.

By adding `#include <sys/wait.h>` at the top of the file, the compiler will be able to recognize and use the `waitpid` function correctly, resolving the error.